fix(#6): improve the code style and add a BranchController
