+++ backend/src/main/java/clf/integra/backend/controller/TestController.java
@@ -36,25 +36,4 @@ public class TestController {
         userService.addUserWithName("A", "B", "C");
         return userRepository.getAllUsers().toString();
     }
-
-    /**
-    * Endpoint to collect taxes and fees from a branch
-    * If the branchId is null or the branch has no customers, returns 404 NOT FOUND
-    * If the revenue collected is 0, returns 417 EXPECTATION FAILED
-    * Otherwise, returns 200 OK with the revenue amount
-     */
-    @PostMapping("branches/{branchId}/collect-taxes-and-fees")
-    public ResponseEntity<Double> collectTaxesAndFeesFromBranch(@PathVariable UUID branchId) {
-        try {
-            double revenue = userService.collectTaxesAndFeesFromBranch(branchId);
-            if (revenue == 0) {
-                return ResponseEntity.status(HttpStatus.EXPECTATION_FAILED).body(null);
-            }
-
-            return ResponseEntity.ok(revenue);
-
-        } catch(IllegalArgumentException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
-        }
-    }
 }
