+++ backend/src/main/java/clf/integra/backend/service/UserService.java
@@ -6,7 +6,6 @@ import clf.integra.backend.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
@@ -23,7 +22,11 @@ public class UserService {
         return uuid;
     }
 
-    public List<String> getAllUsersByBranch(UUID branchId) {
+    public Double getUserBalanceById(UUID id) {
+        return userRepository.getUserBalanceById(id);
+    }
+
+    public List<UserDTO> getAllUsersByBranch(UUID branchId) {
         return userRepository.getAllUsers().stream()
                 .filter(user -> branchId.equals(user.getBranchId()))
                 .map(user -> {
@@ -40,30 +43,24 @@ public class UserService {
             throw new IllegalArgumentException("Branch ID can not be null!");
         }
 
-        // We will be able to use the method of get clients for a branch implemented by Madalina
-        // after merging the develop branch with the branch she created.
-        List<User> allUsers = userRepository.getAllUsers();
-        List<User> usersBranch = new ArrayList<>();
-        double revenue = 0;
-
-        for (User user: allUsers) {
-            if (branchId.equals(user.getId())) {
-                usersBranch.add(user);
-            }
-        }
+        List<User> usersBranch = userRepository.getAllUsers().stream()
+                .filter(user -> branchId.equals(user.getBranchId()))
+                .toList();
 
         if (usersBranch.isEmpty()) {
             throw  new IllegalArgumentException("The branch doesnt not have any customer!");
         }
 
+        double revenue = 0;
         for (User user : usersBranch) {
-            double fee = user.getBalance() < 100 ? user.getBalance() * 0.1 : 10;
+            double userBalance = user.getBalance();
+            double fee = calculateFee(userBalance);
+
             if (fee > 0) {
-                user.setBalance(user.getBalance() - fee);
+                user.setBalance(userBalance - fee);
                 revenue += fee;
             }
         }
-
         return revenue;
     }
 
@@ -71,7 +68,7 @@ public class UserService {
         return UUID.randomUUID();
     }
 
-    public Double getUserBalanceById(UUID id) {
-        return userRepository.getUserBalanceById(id);
+    public double calculateFee(double balance) {
+        return balance < 100 ? balance * 0.1 : 10;
     }
 }
